% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit_data.R
\name{credit}
\alias{credit}
\title{German Credit Dataset (Preprocessed)}
\format{
A data frame with 522 and 6 variables:
\describe{
\item{age}{age of the customer [19-75]}
\item{sex}{sex of the customer (female, male)}
\item{saving.accounts}{saving account balance of the customer (little, moderate, rich)}
\item{duration}{payback duration of credit (in month) [6-72]}
\item{credit.amount}{credit amount [276-18424]}
\item{risk}{whether the credit is of low/good or high/bad risk (bad, good)}
}
}
\description{
Preprocessed version of the German Credit Risk dataset available on kaggle,
based on the Statlog (German credit dataset) of Hofmann (1994) available on
UCI.
}
\details{
The dataset was further adapted: rows with missing values were remove,
low-cardinal classes were binned, classes of the job feature were renamed,
the features on the savings and checking account were defined as ordinal variables,
and all feature names were transposed to lower. Only a subset of features was
selected: "age", "sex", "saving.accounts", "duration", "credit.amount", "risk".
}
\references{
Hofmann, Hans (1994).
\dQuote{Statlog (German Credit Data).}
\emph{UCI Machine Learning Repository}.
\url{https://doi.org/10.24432/C5NC77}.

Ferreira L (2018).
\dQuote{German credit risk.}
Last accessed 10.04.2024, \url{https://www.kaggle.com/datasets/kabure/german-credit-data-with-risk}.
}
