% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3summary-package}
\alias{mlr3summary}
\alias{mlr3summary-package}
\title{mlr3summary: Model and learner summaries for 'mlr3'}
\description{
Summary package for learners and models of the 'mlr3' ecosystem. Returns a summary table that provides an overview of a model at one glance. It works for single models but also resampling results.
}
\author{
\strong{Maintainer}: Susanne Dandl \email{dandls.datascience@gmail.com} (\href{https://orcid.org/0000-0003-4324-4163}{ORCID})

Authors:
\itemize{
  \item Marc Becker \email{marcbecker@posteo.de} (\href{https://orcid.org/0000-0002-8115-0400}{ORCID})
  \item Ludwig Bothmann \email{ludwig.bothmann@stat.uni-muenchen.de} (\href{https://orcid.org/0000-0002-1471-6582}{ORCID})
  \item Giuseppe Casalicchio \email{giuseppe.casalicchio@stat.uni-muenchen.de} (\href{https://orcid.org/0000-0001-5324-5966}{ORCID})
  \item Bernd Bischl \email{bernd_bischl@gmx.net} (\href{https://orcid.org/0000-0001-6002-6980}{ORCID})
}

}
