% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_control}
\alias{summary_control}
\title{Control for model summaries}
\usage{
summary_control(
  measures = NULL,
  importance_measures = importance_choices,
  n_important = 50L
)
}
\arguments{
\item{measures}{(\link[mlr3:Measure]{mlr3::Measure} | list of \link[mlr3:Measure]{mlr3::Measure})\cr
Measure(s) to calculate performance on.}

\item{importance_measures}{(character())\cr
To Do.}

\item{n_important}{(numeric(1))\cr
To Do.}
}
\value{
\link{list}
}
\description{
Various parameters that control aspect of the model summaries.
}
